fix tests after engine update
