+++ test/dependencyReportPanel.test.ts
@@ -3,6 +3,7 @@ import * as chai from 'chai';
 import * as sinon from 'sinon';
 import * as sinonChai from 'sinon-chai';
 
+import { context } from './vscontext.mock';
 import { DependencyReportPanel } from '../src/dependencyReportPanel';
 
 const expect = chai.expect;
@@ -10,32 +11,6 @@ chai.use(sinonChai);
 
 suite('DependencyReportPanel Modules', () => {
   let sandbox: sinon.SinonSandbox;
-
-  let dummyMomentoData = {};
-
-  class DummyMemento implements vscode.Memento {
-    get<T>(key: string): Promise<T | undefined> {
-      return dummyMomentoData[key];
-    }
-    update(key: string, value: any): Promise<any> {
-      dummyMomentoData[key] = value;
-      return Promise.resolve(dummyMomentoData);
-    }
-  }
-
-  const context: vscode.ExtensionContext = {
-    extensionPath: 'path',
-    storagePath: 'string',
-    logPath: 'string',
-    // tslint:disable-next-line:no-empty
-    subscriptions: { dispose(): any { } }[0],
-    workspaceState: new DummyMemento(),
-    globalState: new DummyMemento(),
-    globalStoragePath: 'path',
-    asAbsolutePath(relativePath: string): string {
-      return '';
-    }
-  };
 
   setup(() => {
     sandbox = sinon.createSandbox();
