+++ test/stackAnalysisService.test.ts
@@ -2,39 +2,15 @@ import * as chai from 'chai';
 import * as sinon from 'sinon';
 import * as sinonChai from 'sinon-chai';
 import * as vscode from 'vscode';
-import { stackAnalysisServices } from '../src/stackAnalysisService';
 import * as request from 'request';
 
+import { context } from './vscontext.mock';
+import { stackAnalysisServices } from '../src/stackAnalysisService';
 const expect = chai.expect;
 chai.use(sinonChai);
 
 suite('stacknalysis Services', () => {
   let sandbox: sinon.SinonSandbox;
-  let dummyMomentoData = {};
-
-  class DummyMemento implements vscode.Memento {
-    get<T>(key: string): Promise<T | undefined> {
-      return dummyMomentoData[key];
-    }
-    update(key: string, value: any): Promise<any> {
-      dummyMomentoData[key] = value;
-      return Promise.resolve(dummyMomentoData);
-    }
-  }
-
-  const context: vscode.ExtensionContext = {
-    extensionPath: 'path',
-    storagePath: 'string',
-    logPath: 'string',
-    // tslint:disable-next-line:no-empty
-    subscriptions: { dispose(): any { } }[0],
-    workspaceState: new DummyMemento(),
-    globalState: new DummyMemento(),
-    globalStoragePath: 'path',
-    asAbsolutePath(relativePath: string): string {
-      return '';
-    }
-  };
 
   setup(() => {
     sandbox = sinon.createSandbox();
